requirements.txt
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.8.3
python-dotenv==1.0.0
requests==2.31.0
